+++ account_payment_return_import/models/__init__.py
@@ -1,2 +1,2 @@
-
 from . import payment_return
+from . import payment_return_line
