+++ account_payment_return_import/models/payment_return.py
@@ -1,21 +1,16 @@
-# © 2016 Carlos Dauden <carlos.dauden@tecnativa.com>
-# © 2016 Pedro M. Baeza <pedro.baeza@tecnativa.com>
+# Copyright 2019 ACSONE SA/NV (<https://acsone.eu>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import fields, models
 
 
-class PaymentReturnLine(models.Model):
-    _inherit = 'payment.return.line'
+class PaymentReturn(models.Model):
+    _inherit = 'payment.return'
 
-    # Ensure transactions can be imported only once (if the import format
-    # provides unique transaction ids)
-    unique_import_id = fields.Char('Import ID', readonly=True, copy=False)
-    raw_import_data = fields.Char(
-        help="XML RAW data stored for debugging/check purposes")
-
-    _sql_constraints = [
-        ('unique_import_id',
-         'unique (unique_import_id)',
-         'A payment return transaction can be imported only once!')
-    ]
+    imported_bank_account_id = fields.Many2one(
+        string="Bank account",
+        help="Bank account from the imported file",
+        comodel_name='res.partner.bank',
+        ondelete='restrict',
+        readonly=True,
+    )
